+++ docs/CHANGELOG.md
@@ -32,8 +32,8 @@ Changelog
 - Enh: documentation
 
 ## 0.2.0 (6/6/2023)
-- Enh: black top bar
-- Enh: module compatibility (https://github.com/felixhahnweilheim/humhub-dark-mode/issues/1)
+- Theme: black top bar
+- Theme: module compatibility (https://github.com/felixhahnweilheim/humhub-dark-mode/issues/1)
 
 ## 0.1.0 (30/5/2023)
 - first version